Split neighborhist output into 3 files
The output of neighborhist was a bit unwieldy, especially when there are
a lot of frames in a trajectory. Therefore, the command now
outputs 3 separate files, NN-AllFrames, NN-Combined, and NN-only, all
with appropriate prefix and suffixes add to the filenames.

A function reopen() was also implemented, making it easy to use the same
ofstream variable with multiple files.