Fix segfault when molecule is not defined for atom
When the molecule file does not have a definition for an atom it has an
undefined molecule (NULL), which resulted in a segfault. A message is
now shown when an atom is not assigned to a molecule, and the program
exits cleanly.

When the user requests verbose output, the number of molecules read from
the molecule file is shown.